25.09 fixes

+++ src/plugins/xatta-realtor-admin-bot/server/bot/components/inlineCommands/enterCommand.js
@@ -1,12 +1,4 @@
-const createNewAuthor = require('../../../../botUtils/userController/createNewAuthor');
-
-/**
- * @param user
- * @param localisation
- * @returns {Promise<void>}
- */
-const sendWelcomMessage = async ({ user, localisation }) =>
-    await strapi.bots.admin.sendMessage(user.telegramID, localisation?.WELCOME);
+const { NO_USERNAME } = require('../../../../botUtils/errorHandlers');
 
 /**
  * @param msg
@@ -14,10 +6,9 @@ const sendWelcomMessage = async ({ user, localisation }) =>
  */
 module.exports = async (msg) => {
     const { user, localisation } = msg;
+    if (!process.env.DEVELOPMENT && !msg.user.username) return NO_USERNAME(msg);
     const messageId = msg.message?.message_id || msg.message_id;
-    await strapi.bots.admin.deleteMessage(user.telegramID, messageId);
 
-    await createNewAuthor(msg);
-
-    return sendWelcomMessage({ user, localisation });
+    await strapi.bots.admin.sendMessage(user.telegramID, localisation?.WELCOME);
+    return strapi.bots.admin.deleteMessage(user.telegramID, messageId);
 };
